refactor: Remove deprecated fetch, analyze, and complete commands
Remove deprecated commands completely to simplify codebase:

Breaking Changes:
- Removed 'reviewtask fetch' command (use 'reviewtask' instead)
- Removed 'reviewtask analyze' command (integrated into 'reviewtask')
- Removed 'reviewtask complete' command (use 'reviewtask done' instead)

Updated:
- cmd/root.go: Removed command registrations and legacy command references
- cmd/*_test.go: Updated tests to reflect removed commands
- docs/: Updated all documentation to use new command syntax

Migration Guide:
  reviewtask fetch          →  reviewtask
  reviewtask fetch 123      →  reviewtask 123
  reviewtask analyze        →  reviewtask (auto-included)
  reviewtask complete <id>  →  reviewtask done <id>

Rationale:
- Follows v3.0.0 command simplification goals
- Reduces command count from 15 to 12 (-20%)
- Eliminates confusion between similar commands
- Cleaner codebase with fewer test cases

All tests passing. This completes v3.0.0 command simplification.
