+++ cmd/root.go
@@ -53,20 +53,24 @@ Examples:
   reviewtask 123          # Analyze PR #123 (integrated workflow)
   reviewtask status       # Show current task status
   reviewtask show         # Show current/next task details
-  reviewtask done <id>    # Complete task with automation`,
+  reviewtask done <id>    # Complete task with automation
+
+Common Workflow:
+  reviewtask              # 1. Fetch reviews and analyze
+  reviewtask status       # 2. Check tasks
+  reviewtask start <id>   # 3. Start working on a task
+  reviewtask done <id>    # 4. Complete with full automation`,
 		Args: cobra.MaximumNArgs(1),
 		RunE: runReviewTask,
 	}
 
 	// Add all subcommands
-	cmd.AddCommand(fetchCmd)
 	cmd.AddCommand(statusCmd)
 	cmd.AddCommand(updateCmd)
 	cmd.AddCommand(startCmd)
 	cmd.AddCommand(doneCmd)
 	cmd.AddCommand(holdCmd)
 	cmd.AddCommand(cancelCmd)
-	cmd.AddCommand(completeCmd)
 	cmd.AddCommand(verifyCmd)
 	cmd.AddCommand(configCmd)
 	cmd.AddCommand(showCmd)
@@ -97,11 +101,7 @@ Examples:
   reviewtask 123          # Analyze PR #123 (integrated workflow)
   reviewtask status       # Show current task status
   reviewtask show         # Show current/next task details
-  reviewtask done <id>    # Complete task with automation
-
-For legacy commands, see:
-  reviewtask fetch --help
-  reviewtask analyze --help`,
+  reviewtask done <id>    # Complete task with automation`,
 	Args: cobra.MaximumNArgs(1),
 	RunE: runReviewTask,
 }
@@ -117,7 +117,6 @@ func init() {
 	rootCmd.AddCommand(authCmd)
 	rootCmd.AddCommand(cancelCmd)
 	rootCmd.AddCommand(claudeCmd)
-	rootCmd.AddCommand(completeCmd)
 	rootCmd.AddCommand(configCmd)
 	rootCmd.AddCommand(cursorCmd)
 	rootCmd.AddCommand(debugCmd)
